Update MMM-ImageSlideshow.js
Parameter 'nameonly' code now has modified string size calculation code to remove the "." from the text displayed on screen. For files which are image files, this means the file extension is not displayed i.e. 'image01.jpg' now displays a name of 'image01' as opposed to the previous variant which would have displayed 'image01.'